# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Poseidon API Specification
  description: |-
    A simple API for our financial aggregator app Poseidon.
    Project link:
    - [Poseidon Github Page](https://github.com/Dave-log/Poseidon)
  version: '1.0.0'
servers:
  - url: http://localhost:8080
    description: Local server
...
paths:
...
  /trade/add:
    get:
      summary: Display form for adding a new trade
      tags:
        - Trade
      responses:
        '200':
          description: Add trade form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Add trade form</body></html>"
    post:
      summary: Add a new trade
      tags:
        - Trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trade'
      responses:
        '302':
          description: Trade added successfully and redirected to list page
        '400':
          description: Invalid request, trade not added
        '500':
          description: Internal server error
  /trade/update/{id}:
    get:
      summary: Display form for updating a trade
      tags:
        - Trade
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the trade to update
          schema:
            type: integer
      responses:
        '200':
          description: Update form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Update trade form</body></html>"
    post:
      summary: Update a trade
      tags:
        - Trade
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the trade to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trade'
      responses:
        '302':
          description: Trade updated successfully and redirected to list page
        '400':
          description: Invalid request, trade not updated
        '500':
          description: Internal server error
  /trade/delete/{id}:
    get:
      summary: Delete a trade
      tags:
        - Trade
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the trade to delete
          schema:
            type: integer
      responses:
        '302':
          description: Trade deleted successfully and redirected to list page
        '404':
          description: Trade not found
        '500':
          description: Internal server error
...
  /user/add:
    get:
      summary: Display form for adding a new user
      tags:
        - User
      responses:
        '200':
          description: Add user form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Add user form</body></html>"
    post:
      summary: Add a new user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '302':
          description: User added successfully and redirected to list page
        '400':
          description: Invalid request, user not added
        '500':
          description: Internal server error
  /user/update/{id}:
    get:
      summary: Display form for updating a user
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to update
          schema:
            type: integer
      responses:
        '200':
          description: Update form displayed successfully
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Update user form</body></html>"
    post:
      summary: Update a user
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '302':
          description: User updated successfully and redirected to list page
        '400':
          description: Invalid request, user not updated
        '500':
          description: Internal server error
  /user/delete/{id}:
    get:
      summary: Delete a user
      tags:
        - User
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the user to delete
          schema:
            type: integer
      responses:
        '302':
          description: User deleted successfully and redirected to list page
        '404':
          description: User not found
        '500':
          description: Internal server error
components:
  schemas:
...
    Trade:
      type: object
      properties:
        account:
          type: string
        type:
          type: string
        buyQuantity:
          type: number
      required:
        - account
        - type
        - buyQuantity
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullname:
          type: string
        role:
          type: string
      required:
        - username
        - password
        - fullname
        - role